Refactor + localisation + lot more things
