+++ src/main/java/team5/config/MvcWebConfig.java
@@ -5,8 +5,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-import org.springframework.dao.support.DaoSupport;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import org.springframework.web.servlet.LocaleResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -14,15 +12,11 @@ import org.springframework.web.servlet.config.annotation.ViewResolverRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.i18n.CookieLocaleResolver;
 import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
-import team5.dao.CandidateDAO.CandidateDAO;
-import team5.dao.DBConnector;
-import team5.dao.FeedbackDAO.FeedbackDAO;
-import team5.dao.SkillDao.SkillDao;
-import team5.dao.UserDao.UserDao;
-import team5.dao.SkillDao.SkillDao;
-import team5.dao.vacancyDAO.VacancyDAO;
-
-import java.sql.Connection;
+import team5.dao.CandidateDAO;
+import team5.dao.FeedbackDAO;
+import team5.dao.SkillDao;
+import team5.dao.UserDao;
+import team5.dao.VacancyDAO;
 
 @Configuration
 @EnableWebMvc
