+++ src/main/java/team5/controller/CandidateController.java
@@ -1,7 +1,6 @@
 package team5.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -45,7 +44,7 @@ public class CandidateController {
     public ModelAndView serFeedbackAdd(Locale locale){
         ModelAndView modelAndView = new ModelAndView();
         modelAndView.getModelMap().addAttribute("entity", "Candidate");
-        modelAndView.getModelMap().addAttribute("type", "add");
+        modelAndView.getModelMap().addAttribute("type", "длadd");
         modelAndView.setViewName("index");
         return modelAndView;
     }
