+++ src/main/java/team5/dao/FeedbackDAO.java
@@ -1,6 +1,6 @@
-package team5.dao.FeedbackDAO;
+package team5.dao;
 
-import team5.dao.DBConnector;
+import team5.dao.utils.DBConnector;
 import team5.dao.interfaces.EntityDao;
 import team5.models.Feedback;
