+++ src/main/java/team5/dao/SkillDao.java
@@ -1,7 +1,7 @@
-package team5.dao.SkillDao;
+package team5.dao;
 
 import team5.dao.interfaces.EntityDao;
-import team5.dao.DBConnector;
+import team5.dao.utils.DBConnector;
 import team5.models.Skill;
 
 import java.sql.Connection;
