+++ src/main/java/team5/dao/utils/DBConnector.java
@@ -1,4 +1,4 @@
-package team5.dao;
+package team5.dao.utils;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
