+++ src/main/java/team5/dao/utils/DBUtils.java
@@ -1,4 +1,6 @@
-package team5.dao;
+package team5.dao.utils;
+
+import team5.dao.utils.DBConnector;
 
 import java.sql.Connection;
 import java.sql.Statement;
