+++ src/main/resources/locale/messages.properties
@@ -27,6 +27,20 @@ user.userState=user state
 user.isAdmin=permission
 user.addUserTitle=Adding a new user
 user.addUser=Add user
-user.view=View
-user.add=Add
-sorting.withoutSorting=Without Sorting
+
+feedback.idInterview=Interview ID
+feedback.idInterviewer=Interviewer ID
+feedback.feedbackState=State
+feedback.reason=Reason
+
+button.view=View
+button.add=Add
+skill=Skill
+sort.sortby=Sort by
+sort=Sort
+sort.byname=Name
+sort.byemail=Email
+sort.bysurname=Surname
+sort.none=None
+feedback.addFeedback=ADDING A NEW FEEDBACK
+skill.addSkill=ADDING A NEW SKILL
