+++ src/main/resources/locale/messages_en.properties
@@ -20,13 +20,25 @@ user.userState=user state
 user.isAdmin=permission
 user.addUserTitle=Adding a new user
 user.addUser=Add user
-user.view=View
-user.add=Add
-sorting.withoutSorting=Without Sorting
+button.view=View
+button.add=Add
+sort.sortby=Sort by
 menu.vacancy.id=id
 menu.vacancy.idDeveloper=Developer ID
-munu.vacancy.position=Position
 menu.vacancy.salaryFrom=Salary from
 menu.vacancy.salaryTo=Salary to
 menu.vacancy.vacancyState=State
 menu.vacancy.experienceYearsRequire=Experience years require
+skill=Skill
+feedback.idInterview=Interview ID
+feedback.idInterviewer=Interviewer ID
+feedback.feedbackState=State
+feedback.reason=Reason
+menu.vacancy.position=Position
+sort=Sort
+sort.byname=Name
+sort.byemail=Email
+sort.bysurname=Surname
+sort.none=None
+feedback.addFeedback=ADDING A NEW FEEDBACK
+skill.addSkill=ADDING A NEW SKILL
