+++ src/main/resources/locale/messages_ru.properties
@@ -12,7 +12,7 @@ menu.skill=Навык
 menu.feedback=Отзыв
 menu.vacancy=Вакансия
 team.name=Команда 5
-sorting.withoutSorting=Без сортировки
+sort.sortby=Сортировать по
 user.email=почта
 user.password=пароль
 user.name=имя
@@ -21,12 +21,24 @@ user.userState=статус
 user.isAdmin=права
 user.addUserTitle=Добаление нового пользователя
 user.addUser=Добавить пользователя
-user.view=Просмотр
-user.add=Добавление
+button.view=Просмотр
+button.add=Добавление
 menu.vacancy.id=id
 menu.vacancy.idDeveloper=ID разработчика
-munu.vacancy.position=Позиция
 menu.vacancy.salaryFrom=Зарплата от
 menu.vacancy.salaryTo=Зарплата до
 menu.vacancy.vacancyState=Статус
 menu.vacancy.experienceYearsRequire=Необходимо лет опыта
+skill=Навык
+feedback.idInterview=ID Собеседования
+feedback.idInterviewer=ID собеседующего
+menu.vacancy.position=Позиция
+feedback.reason=Причина
+feedback.feedbackState=Статус
+sort=Сортировать
+sort.byname=Имени
+sort.byemail=Email
+sort.bysurname=Фамилии
+sort.none=Никак
+feedback.addFeedback=ДОБАВЛЕНИЕ НОВОГО ОТЗЫВА
+skill.addSkill=ДОБАВЛЕНИЕ НОВОГО НАВЫКА
