+++ src/main/webapp/WEB-INF/pages/Candidate.jsp
@@ -3,19 +3,19 @@
 <html>
 <head></head>
 <style>
-    <%@include file="../styles/user.css"%>
+    <%@include file="../styles/item.css"%>
 </style>
 <body>
-<div class="user">
-    <div class="user-type">
-        <a href="/candidate/view/1" class="user-type-link">View</a>
-        <a href="/candidate/add" class="user-type-link">Add</a>
+<div class="item">
+    <div class="item-type">
+        <a href="/candidate/view/1" class="item-type-link"><spring:message code="button.view"/></a>
+        <a href="/candidate/add" class="item-type-link"><spring:message code="button.add"/></a>
     </div>
 </div>
 <div>
     <c:if test="${type eq 'view'}">
-        <div class="user-view">
-            <table class="user-table">
+        <div class="item-view">
+            <table class="item-table">
                 <thead>
                 <tr>
                     <th><spring:message code="menu.candidate.id"/></th>
@@ -26,7 +26,7 @@
                 </tr>
                 </thead>
                 <c:forEach var="candidate" items="${candidates}">
-                    <tr class="user-item">
+                    <tr>
                         <td>${candidate.id}</td>
                         <td>${candidate.name}</td>
                         <td>${candidate.surname}</td>
@@ -35,24 +35,24 @@
                     </tr>
                 </c:forEach>
             </table>
-            <div class="user-pages">
+            <div class="item-pages">
                 <c:forEach var="pages" items="${pages}">
-                    <a class="user-page-link" href="/candidate/view/${pages}">${pages}</a>
+                    <a class="item-page-link" href="/candidate/view/${pages}">${pages}</a>
                 </c:forEach>
             </div>
         </div>
     </c:if>
     <c:if test="${type eq 'add'}">
-        <div class="user-add">
-            <form class="user-add-form" method="post" action="/candidateAdd">
-                <label>name</label>
-                <input>
-                <label>surname</label>
-                <input>
-                <label>birthday</label>
-                <input>
-                <label>salary</label>
-                <input>
+        <div class="item-add">
+            <form class="item-add-form" method="post" action="/candidate/add">
+                <label class="item-add-form-label"><spring:message code="menu.candidate.name"/></label>
+                <input class="item-add-form-input">
+                <label class="item-add-form-label"><spring:message code="menu.candidate.surname"/></label>
+                <input class="item-add-form-input">
+                <label class="item-add-form-label"><spring:message code="menu.candidate.birthday"/></label>
+                <input class="item-add-form-input">
+                <label class="item-add-form-label"><spring:message code="menu.candidate.salary"/></label>
+                <input class="item-add-form-input">
                 <button type="submit">submit</button>
             </form>
         </div>
