+++ src/main/webapp/WEB-INF/pages/Vacancy.jsp
@@ -3,19 +3,19 @@
 <html>
 <head></head>
 <style>
-    <%@include file="../styles/user.css"%>
+    <%@include file="../styles/item.css"%>
 </style>
 <body>
-<div class="user">
-    <div class="user-type">
-        <a href="/vacancy/view/1" class="user-type-link">View</a>
-        <a href="/vacancy/add" class="user-type-link">Add</a>
+<div class="item">
+    <div class="item-type">
+        <a href="/vacancy/view/1" class="item-type-link"><spring:message code="button.view"/></a>
+        <a href="/vacancy/add" class="item-type-link"><spring:message code="button.add"/></a>
     </div>
 </div>
 <div>
     <c:if test="${type eq 'view'}">
-        <div class="user-view">
-            <table class="user-table">
+        <div class="item-view">
+            <table class="item-table">
                 <thead>
                 <tr>
                     <th><spring:message code="menu.vacancy.id"/></th>
@@ -28,7 +28,7 @@
                 </tr>
                 </thead>
                 <c:forEach var="vacancy" items="${vacancies}">
-                    <tr class="user-item">
+                    <tr>
                         <td>${vacancy.id}</td>
                         <td>${vacancy.idDeveloper}</td>
                         <td>${vacancy.position}</td>
@@ -39,29 +39,28 @@
                     </tr>
                 </c:forEach>
             </table>
-            <div class="user-pages">
+            <div class="item-pages">
                 <c:forEach var="pages" items="${pages}">
-                    <a class="user-page-link" href="/vacancy/view/${pages}">${pages}</a>
+                    <a class="item-page-link" href="/vacancy/view/${pages}">${pages}</a>
                 </c:forEach>
             </div>
         </div>
     </c:if>
     <c:if test="${type eq 'add'}">
-        <div class="user-add">
-            <form class="user-add-form" method="post" action="/vacancyAdd">
-                <label>idDeveloper</label>
-                <input>
-                <label>Position</label>
-                <input>
-                <label>Salary From</label>
-                <input>
-                <label>Salary To</label>
-                <input>
-                <label>State</label>
-                <input>
-                <label>Experience Years Require</label>
-                <input>
-
+        <div class="item-add">
+            <form class="item-add-form" method="post" action="/vacancyadd">
+                <label class="item-add-form-label"><spring:message code="menu.vacancy.idDeveloper"/></label>
+                <input class="item-add-form-input">
+                <label class="item-add-form-label"><spring:message code="menu.vacancy.position"/></label>
+                <input class="item-add-form-input">
+                <label class="item-add-form-label"><spring:message code="menu.vacancy.salaryFrom"/></label>
+                <input class="item-add-form-input">
+                <label class="item-add-form-label"><spring:message code="menu.vacancy.salaryTo"/></label>
+                <input class="item-add-form-input">
+                <label class="item-add-form-label"><spring:message code="menu.vacancy.vacancyState"/></label>
+                <input class="item-add-form-input">
+                <label class="item-add-form-label"><spring:message code="menu.vacancy.experienceYearsRequire"/></label>
+                <input class="item-add-form-input">
                 <button type="submit">submit</button>
             </form>
         </div>
