--- src/main/webapp/WEB-INF/styles/candidate.css
@@ -1,63 +1,0 @@
-.candidate{
-    display: flex;
-    width: 70%;
-    justify-content: center;
-}
-.candidate-type{
-    display: flex;
-    flex-direction: row;
-}
-.candidate-type-link{
-    margin: 0px 5px;
-    padding: 5px 10px;
-    color: black;
-    text-decoration: none;
-    border-bottom: 1px solid grey;
-    border-top: 1px solid white;
-}
-.candidate-type-link:hover{
-    border-bottom: 2px solid blue;
-    border-top: 0 solid white;
-}
-
-.candidate-view{
-    margin-top: 10px;
-    width: 60%;
-    display: flex;
-    flex-direction: column;
-}
-table.candidate-table {
-    font-family: verdana, arial, sans-serif;
-    font-size: 11px;
-    color: #333333;
-    border-width: 1px;
-    border-color: #3A3A3A;
-    border-collapse: collapse;
-}
-
-table.candidate-table th {
-    border: 1px solid #517994;
-    padding: 8px;
-    background-color: #B2CFD8;
-}
-
-table.candidate-table tr:hover td {
-    background-color: #DFEBF1;
-}
-
-table.candidate-table td {
-    border: 1px solid #517994;
-    padding: 8px;
-    background-color: #ffffff;
-}
-
-.candidate-add{
-    width: 40%;
-    display: flex;
-    justify-content: center;
-    margin-top: 10px;
-}
-.candidate-add-form{
-    display: flex;
-    flex-direction: column;
-}
