+++ src/main/webapp/WEB-INF/styles/item.css
@@ -1,13 +1,13 @@
-.user{
+.item{
     display: flex;
     width: 70%;
     justify-content: center;
 }
-.user-type{
+.item-type{
     display: flex;
     flex-direction: row;
 }
-.user-type-link{
+.item-type-link{
     margin: 0px 5px;
     padding: 5px 10px;
     color: black;
@@ -15,18 +15,18 @@
     border-bottom: 1px solid grey;
     border-top: 1px solid white;
 }
-.user-type-link:hover{
+.item-type-link:hover{
     border-bottom: 2px solid blue;
     border-top: 0 solid white;
 }
 
-.user-view{
+.item-view{
     margin-top: 10px;
     width: 60%;
     display: flex;
     flex-direction: column;
 }
-table.user-table {
+table.item-table {
     font-family: verdana, arial, sans-serif;
     font-size: 11px;
     color: #333333;
@@ -35,28 +35,28 @@ table.user-table {
     border-collapse: collapse;
 }
 
-table.user-table th {
+table.item-table th {
     border: 1px solid #517994;
     padding: 8px;
     background-color: #B2CFD8;
 }
 
-table.user-table tr:hover td {
+table.item-table tr:hover td {
     background-color: #DFEBF1;
 }
 
-table.user-table td {
+table.item-table td {
     border: 1px solid #517994;
     padding: 8px;
     background-color: #ffffff;
 }
-.user-pages{
+.item-pages{
     display: flex;
     justify-content: center;
     align-items: center;
     margin-top: 10px;
 }
-.user-page-link{
+.item-page-link{
     text-decoration: none;
     color: black;
     margin: 0 5px;
@@ -64,16 +64,16 @@ table.user-table td {
     border: 1px #97ffff solid;
     border-radius: 5px;
 }
-.user-page-link:hover{
+.item-page-link:hover{
     background-color: #97FFFF;
 }
 
-.user-add{
+.item-add{
     width: 30%;
     display: flex;
     justify-content: center;
 }
-.user-add-form{
+.item-add-form{
     position: relative;
     display: inline-block;
     max-width: 700px;
@@ -98,7 +98,7 @@ h1 {
     padding: 10px 0;
 }
 
-.user-add-form-input{
+.item-add-form-input{
     -webkit-appearance: none;
     border: 1px solid #ff4a56;
     line-height: 0;
@@ -115,10 +115,10 @@ h1 {
     z-index: 1;
 }
 
-.user-add-form-input, .user-add-form-label {
+.item-add-form-input, .item-add-form-label {
     transition: all 0.25s cubic-bezier(0.53, 0.01, 0.35, 1.5);
 }
-.user-add-form-label {
+.item-add-form-label {
     transform-origin: left center;
     color: #ff4a56;
     font-weight: 100;
@@ -133,26 +133,26 @@ h1 {
     pointer-events: none;
 }
 
-input.user-add-form-input:focus{
+input.item-add-form-input:focus{
     outline: none;
     background: #ff4a56;
     color: white;
     margin-top: 20px;
 }
-input.user-add-form-input:valid{
+input.item-add-form-input:valid{
     margin-top: 25px;
 }
-input.user-add-form-input:focus~label{
+input.item-add-form-input:focus~label{
     transform: translateY(-35px);
     -webkit-transform: translateY(-35px);
 }
 
-input.user-add-form-input:valid~label{
+input.item-add-form-input:valid~label{
     text-transform: uppercase;
     font-style: italic;
     transform: translate(5px, -35px) scale(0.6);
 }
-.user-add-form button{
+.item-add-form button{
     background-color: rgba(255, 74, 86, 0.07);
     margin-top: 10px;
     cursor: pointer;
@@ -168,7 +168,7 @@ input.user-add-form-input:valid~label{
     position: relative;
     z-index: 1;
 }
-.user-add-form button:hover, .user-add-form button:focus{
+.item-add-form button:hover, .item-add-form button:focus{
     color: white;
     background-color: #ff4a56;
 }
